+++ dttools/src/file_cache.c
@@ -49,7 +49,7 @@ static void txn_name( struct file_cache *c, const char *path, char *txn )
 	char shortname[DOMAIN_NAME_MAX];
 	domain_name_cache_guess_short(shortname);
 	md5_buffer(path,strlen(path),digest);
-	sprintf(txn,"%s/txn/%s.%s.%d.XXXXXX",c->root,md5_string(digest),shortname,getpid());
+	sprintf(txn,"%s/txn/%s.%s.%d.XXXXXX",c->root,md5_string(digest),shortname,(int)getpid());
 }
 
 static int wait_for_running_txn( struct file_cache *c, const char *path )
